unifico y creo logica en el service

+++ 3. Spring/ObtenerDiploma/src/main/java/com/ejercicio/ejercicio/models/Diploma.java
@@ -7,7 +7,7 @@ import lombok.Setter;
 @Setter
 public class Diploma {
     protected String mensaje;
-    protected int promedio;
+    protected Double promedio;
     protected String alumno;
 
 }
